travel_ai/fronted/requirements.txt
streamlit
sseclient-py
requests
watchdog
fpdf